requirements.txt
glcontext==3.0.0
moderngl==5.12.0
numpy==2.3.3
pyglet==2.1.9
pyglm==2.8.2
